requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.1.4; python_version >= '3.10'
django-cors-headers==4.6.0; python_version >= '3.9'
django-redis==5.4.0; python_version >= '3.6'
django-rest-framework==0.1.0
djangorestframework==3.15.2; python_version >= '3.8'
dnspython==2.7.0; python_version >= '3.9'
gunicorn==23.0.0; python_version >= '3.7'
numpy==2.2.0; python_version >= '3.12'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pyarrow==18.0.0; python_version >= '3.9'
pyjwt==2.10.1; python_version >= '3.9'
pymongo==4.10.1; python_version >= '3.8'
pymongoarrow==1.6.2; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.2
redis==5.2.1; python_version >= '3.8'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.3; python_version >= '3.8'
tzdata==2024.2; python_version >= '2'
